tests/wpt/tests/css/css-ui/parsing/outline-color-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 3: getComputedStyle().outlineColor</title>
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#outline-color">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #target {
    color: blue;
  }
</style>
</head>
<body>
<div id="target"></div>
<script>
'use strict';
const currentColor = "rgb(0, 0, 255)";
test_computed_value("outline-color", "currentColor", currentColor);
test_computed_value("outline-color", "red", "rgb(255, 0, 0)");

test(() => {
  const target = document.getElementById('target');
  target.style.outlineColor = 'invert';
  if (target.style.outlineColor === 'invert')
    assert_equals(getComputedStyle(target).outlineColor, 'invert');
}, 'invert, if supported, computes to invert');
</script>
</body>
</html>
